Change shininess to accept a float. Simulating emerald from book.

+++ LearnOpenGL/Shapes.hpp
@@ -28,7 +28,7 @@ struct Material
     glm::vec3 ambient = glm::vec3(1.f);
     glm::vec3 diffuse = glm::vec3(1.f);
     glm::vec3 specular = glm::vec3(1.f, 0.f, 0.f);
-    int shininess = 1;
+    float shininess = 1;
 };
