+++ LearnOpenGL/shaders/Phong.frag
@@ -11,7 +11,7 @@ struct Material
     vec3 ambient;
     vec3 diffuse;
     vec3 specular;
-    int shininess;
+    float shininess;
 };
 
 struct Light
